+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/view/dataset_editing/DatasetsOverviewActivity.kt
@@ -21,10 +21,4 @@ class DatasetsOverviewActivity : AppCompatActivity() {
                 .commit()
         }
     }
-
-    /*
-    fun launchDataCreationActivity(view: View){
-        fragment.launchDataCreationActivity(view)
-    }
-     */
 }
