+++ app/src/main/res/layout/fragment_display_dataset.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".view.dataset_editing.DisplayDatasetActivity">
+
+    <EditText
+        android:id="@+id/display_dataset_name"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:inputType="text"
+        android:maxLines="1"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:hint="@string/dataset_name"
+        android:autofillHints="Dataset Name" />
+
+
+    <GridView
+        android:id="@+id/display_dataset_imagesGridView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
+        android:horizontalSpacing="10dp"
+        android:numColumns="2"
+        android:paddingBottom="70dp"
+        android:verticalSpacing="10dp"
+        app:layout_constraintTop_toBottomOf="@id/display_dataset_name" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
+
+
